requirements.txt
robin-stocks==3.4.0
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
finnhub-python==2.4.18

